Move proving FFI into a module
